renderer: compact Vulkan vertex layout 48->32 bytes (M3)
Pack interleaved vertices interleaved at 32 bytes: world pos/normal stay
R32G32B32_SFLOAT (offset 0/12), diffuse color quantized to R8G8B8A8_UNORM
(offset 24), texcoord to R16G16_SFLOAT half-float (offset 28). Matching
vendor-neutral floatToHalf() encoder, RGBA->UNORM quantizer, and updated
VkVertexInputAttributeDescription formats/offsets in the visual binding.

The ambient per-vertex fetch drops from 48 to 32 bytes, reducing vertex
bandwidth for large meshes while keeping f32 precision on positions/normals
needed for correct CAD lighting. Shaders are untouched (SPIR-V inputs are
format-agnostic; the vertex-input description drives decoding). uploadScratch
becomes a byte vector sized VULKAN_VERTEX_STRIDE per vertex.

Verified: vulkan-backend-vertex-color-test, -texture-test, and
-texture-models-test pass with FC_VULKAN_MEM_POOL both 0 and 1 (pool
interacts here via shared geometry blocks). Wide-line path keeps its own
36-byte clip-space layout unchanged.

+++ src/rendering/SoVulkanRenderBackend.h
@@ -891,8 +891,9 @@ private:
 
   // Reusable scratch packing buffer for uploadGeometry().  Resized but never
   // reallocated across successive uploads, so the interleaved repack does not
-  // churn the heap on every geometry change.
-  std::vector<float> uploadScratch;
+  // churn the heap on every geometry change.  Stores the 32-byte packed layout
+  // (see VULKAN_VERTEX_STRIDE), so it is byte-addressed.
+  std::vector<uint8_t> uploadScratch;
   // Reusable per-draw "needs geometry upload" flag buffer for
   // updateGeometryCache().  Grows to the largest draw list seen, then keeps
   // its capacity so an ordinary frame does not heap-allocate a fresh vector
